[refactor] AutoComplete into Lookup
