+++ demos/lookup.php
@@ -6,7 +6,7 @@ require_once __DIR__ . '/database.php';
 // create header
 \atk4\ui\Header::addTo($app, ['Database-driven form with an enjoyable layout']);
 
-\atk4\ui\FormField\AutoComplete::addTo($app, ['placeholder' => 'Search users', 'label' => 'http://'])->setModel(new Country($app->db));
+\atk4\ui\FormField\Lookup::addTo($app, ['placeholder' => 'Search users', 'label' => 'http://'])->setModel(new Country($app->db));
 
 // create form
 $form = \atk4\ui\Form::addTo($app, ['segment']);
@@ -14,19 +14,19 @@ $form = \atk4\ui\Form::addTo($app, ['segment']);
 
 $m = new \atk4\data\Model($db, 'test');
 
-// Without AutoComplete
+// Without Lookup
 $m->hasOne('country1', new Country());
 
-// With AutoComplete
+// With Lookup
 $m->hasOne('country2', [new Country(), 'ui' => ['form' => [
-    'AutoComplete',
+    'Lookup',
     'plus' => true,
 ]]]);
 
 $form->setModel($m);
 
 $form->addField('country3', [
-    'AutoComplete',
+    'Lookup',
     'model'       => new Country($db),
     'placeholder' => 'Search for country by code, LV or UK',
     'search'      => ['name', 'iso', 'iso3'],
@@ -44,13 +44,13 @@ $form->onSubmit(function ($f) use ($db) {
 \atk4\ui\Header::addTo($app, ['Labels']);
 
 // from seed
-\atk4\ui\FormField\AutoComplete::addTo($app, ['placeholder' => 'Search users', 'label' => 'http://'])->setModel(new Country($app->db));
+\atk4\ui\FormField\Lookup::addTo($app, ['placeholder' => 'Search users', 'label' => 'http://'])->setModel(new Country($app->db));
 
 // through constructor
-\atk4\ui\FormField\AutoComplete::addTo($app, ['placeholder' => 'Weight', 'labelRight' => new \atk4\ui\Label(['kg', 'basic'])]);
-\atk4\ui\FormField\AutoComplete::addTo($app, ['label' => '$', 'labelRight' => new \atk4\ui\Label(['.00', 'basic'])]);
+\atk4\ui\FormField\Lookup::addTo($app, ['placeholder' => 'Weight', 'labelRight' => new \atk4\ui\Label(['kg', 'basic'])]);
+\atk4\ui\FormField\Lookup::addTo($app, ['label' => '$', 'labelRight' => new \atk4\ui\Label(['.00', 'basic'])]);
 
-\atk4\ui\FormField\AutoComplete::addTo($app, [
+\atk4\ui\FormField\Lookup::addTo($app, [
     'iconLeft'   => 'tags',
     'labelRight' => new \atk4\ui\Label(['Add Tag', 'tag']),
 ]);
@@ -60,46 +60,19 @@ $label = new \atk4\ui\Label();
 $label->addClass('left corner');
 \atk4\ui\Icon::addTo($label, ['asterisk']);
 
-\atk4\ui\FormField\AutoComplete::addTo($app, [
+\atk4\ui\FormField\Lookup::addTo($app, [
     'label' => $label,
 ])->addClass('left corner');
 
 \atk4\ui\Header::addTo($app, ['Auto-complete inside modal']);
 
 $modal = \atk4\ui\Modal::addTo($app)->set(function ($p) {
-    $a = \atk4\ui\FormField\AutoComplete::addTo($p, ['placeholder' => 'Search users', 'label' => 'http://']);
+    $a = \atk4\ui\FormField\Lookup::addTo($p, ['placeholder' => 'Search users', 'label' => 'http://']);
     $a->setModel(new Country($p->app->db));
 });
-\atk4\ui\Button::addTo($app, ['Open autocomplete on a Modal window'])->on('click', $modal->show());
+\atk4\ui\Button::addTo($app, ['Open Lookup on a Modal window'])->on('click', $modal->show());
 
-\atk4\ui\Header::addTo($app, ['New Lookup field']);
-
-$form = \atk4\ui\Form::addTo($app, ['segment']);
-\atk4\ui\Label::addTo($form, ['Input new country information here', 'top attached'], ['AboveFields']);
-
-$c = new Country($db);
-$c->addExpression('letter1', 'concat("Ends with ", substring([name], -1))');
-
-$form->addField('country_a', [
-    'Lookup',
-    'model'       => new Country($db),
-    'hint'        => 'Lookup field is just like AutoComplete, supports all the same options.',
-    'placeholder' => 'Search for country by code, LV or UK',
-    'search'      => ['name', 'iso', 'iso3'],
-]);
-
-$lookup = $form->addField('country_b', [
-    'Lookup',
-    'model'       => $c,
-    'hint'        => 'However one or few "filtering" options can be added narrowing down the final result set',
-    'placeholder' => 'Search for country by code, LV or UK',
-    'search'      => ['name', 'iso', 'iso3'],
-]);
-$lookup->addFilter('letter1');
-
-$form->buttonSave->set('Add Countries');
-
-\atk4\ui\Header::addTo($app, ['Auto-complete dependency']);
+\atk4\ui\Header::addTo($app, ['Lookup dependency']);
 
 $form = \atk4\ui\Form::addTo($app, ['segment']);
 \atk4\ui\Label::addTo($form, ['Input information here', 'top attached'], ['AboveFields']);
@@ -123,7 +96,7 @@ $form->addField('contains', [
 ]);
 
 $lookup = $form->addField('country', [
-    'AutoComplete',
+    'Lookup',
     'model'       => new Country($db),
     'dependency'  => function ($model, $data) {
         $conditions = [];
@@ -162,7 +135,7 @@ $form->addField('ends_with', [
 ]);
 
 $lookup = $form->addField('country', [
-    'AutoComplete',
+    'Lookup',
     'model'       => new Country($db),
     'dependency'  => function ($model, $data) {
         isset($data['ends_with']) ? $model->addCondition('name', 'like', '%'.$data['ends_with']) : null;
