+++ src/FormField/Lookup.php
@@ -12,6 +12,13 @@ class Lookup extends Input
     public $ui = 'input';
 
     /**
+     * Declare this property so Lookup is consistent as decorator to replace FormField\DropDown.
+     *
+     * @var array
+     */
+    public $values = [];
+
+    /**
      * Object used to capture requests from the browser.
      *
      * @var callable
@@ -31,18 +38,29 @@ class Lookup extends Input
      *
      * If left null, then search will be performed on a model's title field
      *
-     * @var array|Closure
+     * @var array|\Closure
      */
     public $search;
+
+    /**
+     * If a dependency callback is declared Lookup collects the current (dirty) form values
+     * and passes them on to the dependency callback so conditions on the field model can be applied.
+     * This allows for generating different option lists depending on dirty form values
+     * E.g if we have a dropdown field 'country' we can add to the form an Lookup field 'state'
+     * with dependency
+     * Then model of the 'state' field can be limited to states of the currently selected 'country'.
+     *
+     * @var callable
+     */
+    public $dependency;
 
     /**
      * Set this to create right-aligned button for adding a new a new record.
      *
      * true = will use "Add new" label
      * string = will use your string
-     * array = will use your array as defaults for button
      *
-     * @var null|bool|string|array
+     * @var null|bool|string
      */
     public $plus = false;
 
@@ -80,11 +98,11 @@ class Lookup extends Input
     /**
      * Semantic UI dropdown module settings.
      * Use this setting to configure various dropdown module settings
-     * to use with Autocomplete.
+     * to use with Lookup.
      *
      * For example, using this setting will automatically submit
      * form when field value is changes.
-     * $form->addField('field', ['AutoComplete', 'settings'=>['allowReselection' => true,
+     * $form->addField('field', ['Lookup', 'settings'=>['allowReselection' => true,
      *                           'selectOnKeydown' => false,
      *                           'onChange'        => new atk4\ui\jsExpression('function(value,t,c){
      *                                                          if ($(this).data("value") !== value) {
@@ -99,69 +117,32 @@ class Lookup extends Input
     public $settings = [];
 
     /**
-     * Default string for presenting filter.
+     * Define callback for generating the row data
+     * If left empty default callback Lookup::defaultRenderRow is used.
      *
-     * @var string
+     * @var null|callable
      */
-    public $filterHeaderLabel = 'Filtering options:';
+    public $renderRowFunction;
 
     /**
-     * Default label when no data is selected in filter.
+     * Whether or not to accept multiple value.
+     *   Multiple values are sent using a string with comma as value delimiter.
+     *   ex: 'value1,value2,value3'.
      *
-     * @var string
+     * @var bool
      */
-    public $filterEmpty = 'All';
-
-    public $filterChain = null;
-
-    /**
-     * Array containing filters.
-     *
-     * @var null|array
-     */
-    public $filters = null;
+    public $multiple = false;
 
     public function init()
     {
         parent::init();
 
-        $this->label = null;
+        $this->template->set([
+            'input_id'    => $this->name.'-ac',
+            'placeholder' => $this->placeholder,
+        ]);
 
-        $this->template->set('input_id', $this->name.'-ac');
-
-        $this->template->set('place_holder', $this->placeholder);
-
-        if ($this->plus) {
-            $this->action = $this->factory(['Button', is_array($this->plus) || is_string($this->plus) ? $this->plus : 'Add new', 'disabled' => ($this->disabled || $this->readonly)], null, 'atk4\ui');
-        }
-        //var_Dump($this->model->get());
-        if ($this->form) {
-            $vp = \atk4\ui\VirtualPage::addTo($this->form);
-        } else {
-            $vp = \atk4\ui\VirtualPage::addTo($this->owner);
-        }
-
-        $vp->set(function ($p) {
-            $f = \atk4\ui\Form::addTo($p);
-            $f->setModel($this->model);
-
-            $f->onSubmit(function ($f) {
-                $id = $f->model->save()->id;
-
-                $modal_chain = new jQuery('.atk-modal');
-                $modal_chain->modal('hide');
-                $ac_chain = new jQuery('#'.$this->name.'-ac');
-                $ac_chain->dropdown('set value', $id)->dropdown('set text', $f->model->getTitle());
-
-                return [
-                    $modal_chain,
-                    $ac_chain,
-                ];
-            });
-        });
-        if ($this->action) {
-            $this->action->js('click', new \atk4\ui\jsModal('Adding New Record', $vp));
-        }
+        $this->initQuickNewRecord();
 
         $this->settings['forceSelection'] = false;
     }
@@ -174,121 +155,193 @@ class Lookup extends Input
         return $this->callback->getJSURL();
     }
 
-    public function getData()
+    /**
+     * Generate API response.
+     */
+    public function outputApiResponse()
     {
-        if (!$this->model) {
-            $this->app->terminate(json_encode([['id' => '-1', 'name' => 'Model must be set for AutoComplete']]));
-        }
-
-        $id_field = $this->id_field ?: $this->model->id_field;
-        $title_field = $this->title_field ?: $this->model->title_field;
-
-        $this->renderFilters();
-        $this->applyFilters();
-
-        $this->model->setLimit($this->limit);
-
-        if (isset($_GET['q'])) {
-            if ($this->search instanceof Closure) {
-                $this->search($this->model, $_GET['q']);
-            } elseif ($this->search && is_array($this->search)) {
-                $this->model->addCondition(array_map(function ($field) {
-                    return [$field, 'like', '%'.$_GET['q'].'%'];
-                }, $this->search));
-            } else {
-                $this->model->addCondition($title_field, 'like', '%'.$_GET['q'].'%');
-            }
-        }
-
-        $data = [];
-        foreach ($this->model as $junk) {
-            // IMPORTANT: always convert data to string, otherwise numbers can be rounded by JS
-            $data[] = [
-                'id'   => (string) $this->model[$id_field],
-                'name' => (string) $this->model[$title_field],
-            ];
-        }
-
-        if ($this->empty) {
-            array_unshift($data, ['id' => '0', 'name' => (string) $this->empty]);
-        }
-
         $this->app->terminate(json_encode([
             'success' => true,
-            'results' => $data,
+            'results' => $this->getData(),
         ]));
     }
 
     /**
-     * Add filter dropdown.
+     * Generate Lookup data.
      *
-     * Ex: you need to restreint value available in city dropdown base on user input Country and language
+     * @param bool $limit
      *
-     *      $l = $form->addField('city',['Lookup']);
-     *      $l->addFilter('country_test', 'Country');
-     *      $l->addFilter('language', 'Lang');
-     *
-     * This way, dropdown value will contains city corresponding to proper country and/or language.
-     *
-     * @param $field
-     * @param null $label
+     * @return array
      */
-    public function addFilter($field, $label = null)
+    public function getData($limit = true)
     {
-        if (!$this->model->hasField($field)) {
-            throw new \atk4\ui\Exception([
-                'Unable to filter by non-existant field',
-                'field'=> $field,
-            ]);
+        if (!$this->model) {
+            return [['value' => '-1', 'title' => 'Model must be set for Lookup']];
         }
-        $this->filters[] = ['field' => $field, 'label' => $label];
+
+        $this->applyLimit($limit);
+
+        $this->applySearchConditions();
+
+        $this->applyDependencyConditions();
+
+        $data = [];
+        foreach ($this->model as $row) {
+            $data[] = $this->renderRow($row);
+        }
+
+        if (!$this->multiple && $this->empty) {
+            array_unshift($data, ['value' => '0', 'title' => (string) $this->empty]);
+        }
+
+        return $data;
     }
 
     /**
-     * Check if filtering is need.
+     * Renders the Lookup row depending on properties set.
+     *
+     * @param array $row
+     *
+     * @return mixed
      */
-    public function applyFilters()
+    public function renderRow($row)
     {
-        if ($this->filters) {
-            $requested_filter = $_GET['filter'] ?? null;
-            foreach ($this->filters as $k => $filter) {
-                if (
-                    isset($_GET[$filter['field']]) &&
-                    !empty($_GET[$filter['field']]) &&
-                    $_GET[$filter['field']] != $this->filterEmpty &&
-                    $requested_filter != $filter['field']
-                ) {
-                    $this->model->addCondition($filter['field'], $_GET[$filter['field']]);
+        $renderRowFunction = is_callable($this->renderRowFunction) ? $this->renderRowFunction : [__CLASS__, 'defaultRenderRow'];
+
+        return call_user_func($renderRowFunction, $this, $row);
+    }
+
+    /**
+     * Default callback for generating data row.
+     *
+     * @param Lookup           $field
+     * @param \atk4\data\Model $row
+     * @param string           $key
+     *
+     * @return string[]
+     */
+    public static function defaultRenderRow($field, $row, $key = null)
+    {
+        $id_field = $field->id_field ?: $row->id_field;
+        $title_field = $field->title_field ?: $row->title_field;
+
+        // IMPORTANT: always convert data to string, otherwise numbers can be rounded by JS
+        return [
+            'value' => (string) $row[$id_field],
+            'title' => (string) $row[$title_field],
+        ];
+    }
+
+    /**
+     * Add button for new record.
+     */
+    protected function initQuickNewRecord()
+    {
+        if (!$this->plus) {
+            return;
+        }
+
+        $this->plus = is_bool($this->plus) ? 'Add New' : $this->plus;
+
+        $this->plus = is_string($this->plus) ? ['button' => $this->plus] : $this->plus;
+
+        $buttonSeed = $this->plus['button'] ?? [];
+
+        $buttonSeed = is_string($buttonSeed) ? ['content' => $buttonSeed] : $buttonSeed;
+
+        $defaultSeed = ['Button', 'disabled' => ($this->disabled || $this->readonly)];
+
+        $this->action = $this->factory(array_merge($defaultSeed, (array) $buttonSeed), null, 'atk4\ui');
+
+        if ($this->form) {
+            $vp = \atk4\ui\VirtualPage::addTo($this->form);
+        } else {
+            $vp = \atk4\ui\VirtualPage::addTo($this->owner);
+        }
+
+        $vp->set(function ($page) {
+            $form = \atk4\ui\Form::addTo($page);
+
+            $model = clone $this->model;
+
+            $form->setModel($model->onlyFields($this->plus['fields'] ?? []));
+
+            $form->onSubmit(function ($form) {
+                $form->model->save();
+
+                $ret = [
+                    (new jQuery('.atk-modal'))->modal('hide'),
+                ];
+
+                if ($row = $this->renderRow($form->model)) {
+                    $chain = new jQuery('#'.$this->name.'-ac');
+                    $chain->dropdown('set value', $row['value'])->dropdown('set text', $row['title']);
+
+                    $ret[] = $chain;
                 }
-            }
+
+                return $ret;
+            });
+        });
+
+        $caption = $this->plus['caption'] ?? 'Add New '.$this->model->getModelCaption();
+
+        $this->action->js('click', new \atk4\ui\jsModal($caption, $vp));
+    }
+
+    /**
+     * Apply limit to model.
+     */
+    protected function applyLimit($limit = true)
+    {
+        if (!$limit) {
+            return;
+        }
+
+        $this->model->setLimit(is_numeric($limit) ? $limit : $this->limit);
+    }
+
+    /**
+     * Apply conditions to model based on search string.
+     */
+    protected function applySearchConditions()
+    {
+        if (!isset($_GET['q'])) {
+            return;
+        }
+
+        if ($this->search instanceof \Closure) {
+            $this->search($this->model, $_GET['q']);
+        } elseif ($this->search && is_array($this->search)) {
+            $this->model->addCondition(array_map(function ($field) {
+                return [$field, 'like', '%'.$_GET['q'].'%'];
+            }, $this->search));
+        } else {
+            $title_field = $this->title_field ?: $this->model->title_field;
+
+            $this->model->addCondition($title_field, 'like', '%'.$_GET['q'].'%');
         }
     }
 
     /**
-     * Check if filtering is needed and terminate app execution.
+     * Apply conditions to model based on dependency.
      */
-    public function renderFilters()
+    protected function applyDependencyConditions()
     {
-        if (isset($_GET['filter'])) {
-            if (isset($_GET['q'])) {
-                $this->model->addCondition($_GET['filter'], 'like', '%'.$_GET['q'].'%');
-            }
-            // Apply filtering to filter.
-            $this->applyFilters();
-            $action = $this->model->action('field', [$_GET['filter']]);
-            $action->group($_GET['filter']);
-            $rows = $action->get();
-            $data = [];
-            foreach ($rows as $k => $v) {
-                $data[] = ['id' => $k, 'name' => $v[$_GET['filter']]];
-            }
-            array_unshift($data, ['id' => -1, 'name' => $this->filterEmpty]);
-
-            $this->app->terminate(json_encode([
-                'success' => true,
-                'results' => $data,
-            ]));
+        if (!is_callable($this->dependency)) {
+            return;
         }
+
+        $data = [];
+        if (isset($_GET['form'])) {
+            parse_str($_GET['form'], $data);
+        } elseif ($this->form) {
+            $data = $this->form->model->get();
+        } else {
+            return;
+        }
+
+        call_user_func($this->dependency, $this->model, $data);
     }
 
     /**
@@ -297,24 +350,13 @@ class Lookup extends Input
     public function getInput()
     {
         return $this->app->getTag('input', array_merge([
-            'name'        => $this->short_name,
-            'type'        => 'hidden',
-            'id'          => $this->id.'_input',
-            'value'       => $this->getValue(),
-            'readonly'    => $this->readonly ? 'readonly' : false,
-            'disabled'    => $this->disabled ? 'disabled' : false,
+            'name'     => $this->short_name,
+            'type'     => 'hidden',
+            'id'       => $this->id.'_input',
+            'value'    => $this->getValue(),
+            'readonly' => $this->readonly ? 'readonly' : false,
+            'disabled' => $this->disabled ? 'disabled' : false,
         ], $this->inputAttr));
-    }
-
-    public function getFilterInput($name, $id, $value = null)
-    {
-        return $this->app->getTag('input', [
-            'name' => $name,
-            'type' => 'hidden',
-            /*'placeholder'=> $this->placeholder,*/
-            'id'    => $id,
-            'value' => $value,
-        ]);
     }
 
     /**
@@ -332,136 +374,28 @@ class Lookup extends Input
     }
 
     /**
-     * Override this method if you want to add more logic to the initialization of the
-     * auto-complete field.
+     * Override this method if you want to add more logic to the initialization of the auto-complete field.
      *
      * @param jQuery
      */
     protected function initDropdown($chain)
     {
         $settings = array_merge([
-            'fields'      => ['name' => 'name', 'value' => 'id'/*, 'text' => 'description'*/],
+            'fields'      => ['name' => 'title'],
             'apiSettings' => array_merge(['url' => $this->getCallbackURL().'&q={query}'], $this->apiConfig),
         ], $this->settings);
 
         $chain->dropdown($settings);
     }
 
-    /**
-     * For each filters, we need to create js dropdown and setup proper api settings.
-     *
-     * When changing value on any of the filter we need to set main dropdown
-     * apiSettings again in order for main dropdown to send proper filter value.
-     * This is done via the onChange function of the dropdown filter.
-     *
-     * This function will generate js similar to this:
-     *   $("#filterName1").dropdown({
-     *      "fields":{
-     *                  "name":"name",
-     *                  "value":"id"
-     *                },
-     *      "apiSettings":{
-     *                      "url":"autocomplete.php?atk_admin_form_generic_country2_callback=ajax\x26__atk_callback=1\x26q={query}",
-     *                      "cache":false,
-     *                      "data":{"filter":"city"}
-     *                  },
-     *      "onChange":function() {
-     *           $("#mainDropdown").dropdown(
-     *              {"fields":{"name":"name","value":"id"},
-     *              "apiSettings":{"url":"autocomplete.php?atk_admin_form_generic_country2_callback=ajax\x26__atk_callback=1\x26q={query}",
-     *              "cache":false,
-     *              "data":{"filteName1":$("input[name=filteName1]").parent().find(".text").text(),"filteName2":$("input[name=filteName2]").parent().find(".text").text()}
-     *               }
-     *          });
-     *    }});
-     *
-     * @throws \atk4\ui\Exception
-     */
-    public function createFilterJsDropdown()
-    {
-        foreach ($this->filters as $k => $filter) {
-            $f_name = $this->name.'-ac_f'.$k;
-            $chain = new jQuery('#'.$f_name);
-            $options = [
-                'fields'      => ['name' => 'name', 'value' => 'id'],
-                'match'       => 'value',
-                'apiSettings' => [
-                    'url'        => $this->getCallbackURL().'&q={query}',
-                    'cache'      => false,
-                    'data'       => array_merge($this->getFilterQuery(), ['filter' => $filter['field']]),
-                    'onResponse' => new jsFunction(['resp'], [
-                        new jsExpression('if (!resp.success){atk.apiService.atkSuccessTest(resp);}'),
-                    ]),
-                ],
-                'onChange'   => new jsFunction([
-                    (new jQuery())->trigger('filterChanged'),
-                    $this->getJsDropdown(),
-                ]),
-            ];
-
-            $this->js(true, $chain->dropdown($options));
-        }
-        //set filter value using $(this) context for onChange handler instead of filter name.
-        $options['apiSettings']['data']['filter'] = (new jQuery())->find('input')->attr('name');
-        $this->filterChain = $options;
-    }
-
-    /**
-     * Return the main dropdown js chain.
-     *
-     * @return jQuery
-     */
-    public function getJsDropdown()
-    {
-        $chain = new jQuery('#'.$this->name.'-ac');
-        $this->initDropdown($chain);
-
-        return $chain;
-    }
-
-    /**
-     * Will create jsExpression need to add to the dropdown apiSettings data when using filters.
-     *
-     * ex: {"data":{"filterName1":$("input[name=filterName1]").parent().find(".text").text(),"filterName2":$("input[name=filterName2]").parent().find(".text").text()}}
-     *
-     * @return array
-     */
-    public function getFilterQuery()
-    {
-        $q = [];
-        foreach ($this->filters as $key => $filter) {
-            $q[$filter['field']] = new jsExpression('$([input]).parent().dropdown("get text")', ['input' => 'input[name='.$filter['field'].']']);
-        }
-
-        return $q;
-    }
-
-    /**
-     * When filter changed value,
-     * we need to regenerate each filter dropdown.
-     *
-     * Note: regenerating dropdown seem to have them lost their
-     * text value. Need to reset text value by getting them prior to regenerating dropdown.
-     *
-     * @throws \atk4\ui\Exception
-     *
-     * @return jQuery
-     */
-    public function onJsFilterChanged()
-    {
-        return (new jQuery())
-            ->on('filterChanged',
-                 new jsFunction(array_merge(
-                                    [new jsExpression('let dropdowns = $(".atk-filter-dropdown")')],
-                                    [new jsExpression('dropdowns.each(function(){const value = $(this).dropdown("get text"); $(this).dropdown([chain]); $(this).dropdown("set text", value); $(this).dropdown("refresh");})', ['chain' => $this->filterChain])]
-                                ))
-            );
-    }
-
     public function renderView()
     {
         $this->callback = \atk4\ui\Callback::addTo($this);
-        $this->callback->set([$this, 'getData']);
+        $this->callback->set([$this, 'outputApiResponse']);
+
+        if ($this->multiple) {
+            $this->template->set('multiple', 'multiple');
+        }
 
         if ($this->disabled) {
             $this->settings['showOnFocus'] = false;
@@ -478,57 +412,46 @@ class Lookup extends Input
             $this->template->set('readonly', 'readonly');
         }
 
-        if ($this->filters) {
-            $this->fieldClass = 'ui segment';
-
-            //add filtering query to main dropdown
-            $this->apiConfig = array_merge($this->apiConfig, ['data' => $this->getFilterQuery()]);
-
-            //render filters to template
-            $this->template->set('FilterHeaderLabel', $this->filterHeaderLabel);
-            $ft = $this->template->cloneRegion('Dropdown');
-            $html = '';
-
-            foreach ($this->filters as $k => $filter) {
-                $f_name = $this->name.'-ac_f'.$k;
-                $ft->set('input_id', $f_name);
-                $ft->set('FilterLabel', $filter['label']);
-                $ft->set('place_holder', $this->filterEmpty);
-                $ft->set('filterClass', 'atk-filter-dropdown');
-                $ft->setHTML('Input', $this->getFilterInput($filter['field'], $filter['field'].'_id'));
-                $html .= $ft->render();
-            }
-            $this->template->setHTML('Filters', $html);
-
-            // create proper js for dropdown.
-            $this->createFilterJsDropdown();
-
-            // output changed handler for filters.
-            $this->js(true, $this->onJsFilterChanged());
-        } else {
-            $this->template->del('FilterContainer');
-            $this->template->del('FilterLabelContainer');
+        if ($this->dependency) {
+            $this->apiConfig['data'] = array_merge([
+                'form' => new jsFunction([new jsExpression('return []', [$this->js()->closest('form')->serialize()])]),
+            ], $this->apiConfig['data'] ?? []);
         }
 
-        $this->js(true, $this->getJsDropdown());
+        $chain = new jQuery('#'.$this->name.'-ac');
+
+        $this->initDropdown($chain);
 
         if ($this->field && $this->field->get()) {
             $id_field = $this->id_field ?: $this->model->id_field;
-            $title_field = $this->title_field ?: $this->model->title_field;
 
             $this->model->tryLoadBy($id_field, $this->field->get());
 
-            if (!$this->model->loaded()) {
-                $this->field->set(null);
+            if ($this->model->loaded()) {
+                $row = $this->renderRow($this->model);
+
+                $chain->dropdown('set value', $row['value'])->dropdown('set text', $row['title']);
             } else {
-                $chain = new jQuery('#'.$this->name.'-ac');
-                // IMPORTANT: always convert data to string, otherwise numbers can be rounded by JS
-                $chain->dropdown('set value', (string) $this->model[$id_field])
-                        ->dropdown('set text', (string) $this->model[$title_field]);
-                $this->js(true, $chain);
+                $this->field->set(null);
             }
         }
 
+        $this->js(true, $chain);
+
         parent::renderView();
+    }
+
+    /**
+     * Convert value to expected comma separated list before setting it.
+     *
+     * {@inheritdoc}
+     *
+     * @see \atk4\ui\FormField\Generic::set()
+     */
+    public function set($value = null, $junk = null)
+    {
+        $value = implode(',', (array) $value);
+
+        return parent::set($value, $junk);
     }
 }
